custom_components/dominion_energy/requirements.txt
pyfunctional==1.5.0
selenium==4.25.0
arrow==1.3.0
webdriver-manager==4.0.2
polars==1.13.1
openpyxl==3.1.5
fastexcel==0.12.0